refactor(core): split engine and client modules
